Add casts.
Former-commit-id: f8d46425fec877b7859065a0b47da1b925cd8edb

+++ client/src/gui/widgets/buddy.c
@@ -111,7 +111,7 @@ void widget_buddy_remove(widgetdata *widget, const char *name)
 		uint32 row;
 
 		tmp = WIDGET_BUDDY(widget);
-		utarray_erase(tmp->names, idx, 1);
+		utarray_erase(tmp->names, (size_t) idx, 1);
 
 		for (row = 0; row < tmp->list->rows; row++)
 		{
